Fixed refresh (i hope), added new media choice page and tools page, changed routing
