+++ DjangoSite/landing/templates/landing/index.html
@@ -43,7 +43,7 @@
 							<a href="/resume">About</a>
 						</li>
 						<li>
-							<a href="/media">Media Server Access</a>
+							<a href="/media">Media DB</a>
 						</li>
 						<li>
 							<a href="/admin">Admin</a>
