+++ DjangoSite/media/templates/media/mediaChoice.html
@@ -1,0 +1,113 @@
+{%load static%}
+<!DOCTYPE html>
+<!--[if lt IE 7]>      <html class="no-js lt-ie9 lt-ie8 lt-ie7"> <![endif]-->
+<!--[if IE 7]>         <html class="no-js lt-ie9 lt-ie8"> <![endif]-->
+<!--[if IE 8]>         <html class="no-js lt-ie9"> <![endif]-->
+<!--[if gt IE 8]>      <html class="no-js"> <!--<![endif]-->
+<html
+	data-bs-theme="{{colorMode}}"
+	data-theme="{{colorMode}}">
+	<head>
+		<meta charset="utf-8" />
+		<meta
+			http-equiv="X-UA-Compatible"
+			content="IE=edge" />
+		<title>Developer Tools</title>
+		<meta
+			name="viewport"
+			content="width=device-width, initial-scale=1" />
+		<link
+			href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css"
+			rel="stylesheet" />
+		<meta
+			name="color-scheme"
+			content="light dark" />
+		<link
+			rel="stylesheet"
+			href="https://cdn.jsdelivr.net/npm/@picocss/pico@2.0.6/css/pico.css" />
+		<link
+			rel="stylesheet"
+			href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css" />
+		<link
+			rel="stylesheet"
+			href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.3/font/bootstrap-icons.min.css" />
+		<link
+			rel="stylesheet"
+			href="{% static 'styles/MasterStyle.css' %}" />
+	</head>
+	<body>
+		<main class="container">
+			<header style="height: 5vh">
+				<nav>
+					<ul>
+						<li>
+							<strong>Media Selection</strong>
+						</li>
+					</ul>
+					<ul>
+						<li>
+							<button href="">Home</button>
+						</li>
+						<li>
+							<button href="/media">Media DB</button>
+						</li>
+						<li>
+							<button href="/admin">Admin</button>
+						</li>
+					</ul>
+				</nav>
+			</header>
+			<div
+				class="d-flex flex-column justify-content-center align-items-center gap-3 pt-5 pb-0"
+				style="height: 10vh">
+				<h1>Select Media Type</h1>
+			</div>
+			<div
+				class="d-flex flex-column gap-3 pt-4"
+				style="height: 80vh">
+				<div class="d-flex flex-row justify-content-center align-items-center gap-3">
+					<button
+						style="min-width: 10vw; min-height: 10vh"
+						onclick="document.location = '/media/new?type=movie'">
+						<i class="mx-3 bi bi-film"></i>
+					</button>
+					<button
+						style="min-width: 10vw; min-height: 10vh"
+						onclick="document.location = '/media/new?type=tvshow'">
+						<i class="mx-3 bi bi-tv"></i>
+					</button>
+
+					<button
+						style="min-width: 10vw; min-height: 10vh"
+						onclick="document.location = '/media/new?type=comic'">
+						<i class="mx-3 bi bi-book"></i>
+					</button>
+				</div>
+				<div class="d-flex flex-row justify-content-center align-items-center gap-3">
+					<button
+						style="min-width: 10vw; min-height: 10vh"
+						onclick="document.location = '/media/new?type=novel'">
+						<i class="mx-3 bi bi-journal-bookmark"></i>
+					</button>
+					<button
+						style="min-width: 10vw; min-height: 10vh"
+						onclick="document.location = '/media/new?type=album'">
+						<i class="mx-3 bi bi-vinyl"></i>
+					</button>
+					<button
+						style="min-width: 10vw; min-height: 10vh"
+						onclick="document.location = '/media/new?type=podcast'">
+						<i class="mx-3 bi bi-mic"></i>
+					</button>
+				</div>
+				<div class="d-flex flex-row justify-content-center align-items-center gap-3">
+					<button
+						style="min-width: 10vw; min-height: 10vh"
+						onclick="document.location = '/media/new?type=youtube'">
+						<i class="mx-3 bi bi-youtube"></i>
+					</button>
+				</div>
+			</div>
+		</main>
+	</body>
+</html>
