+++ DjangoSite/media/templates/media/stats.html
@@ -50,7 +50,7 @@
 							<a href="/resume">About</a>
 						</li>
 						<li>
-							<a href="/media">Media Server Access</a>
+							<a href="/media">Media DB</a>
 						</li>
 						<li>
 							<a href="/admin">Admin</a>
